grass.script: Use description parameter in create_project (#6440)
This adds description parameter to create_project as a full word replacing (but keeping) the desc parameter.

The behavior slightly changes so that MYNAME (project description file) is only created when description is provided. This is what the current documentation suggests ('desc...creates MYNAME file'), but it was implemented that MYNAME was always created even if empty.

The description file is now created for empty text (description is empty), but not for None (no description).

This is a project description specific part of #6437.