tests: Use db.test tool in tests
The db.test tool allows to run a basic suite of SQL tests with the given connection in a mapset, but the tool itself was not used in the tests. This adds pytest-based tests which use db.test with SQLite driver.

Additionally, this fixes the code to provide a better message when the connection information (and/or the connection itself) does not exist.