tests: Use db.test tool in tests (#6473)
The db.test tool allows to run a basic suite of SQL tests with the given connection in a mapset, but the tool itself was not used in the tests. This adds pytest-based tests which use db.test with SQLite driver.

Additionally, this fixes the code to provide a better message when the connection information (and/or the connection itself) does not exist.

In the Makefile, specify the db.test SQL test text file explicitly, avoiding wildcard which matches the tests directory. (CMakeList.txt already does that.)

Run db.connect with -c which fits with the documentation, although running without parameters seems to be possible on Linux and macOS (but fails on Windows).