tests: Use copy of env for init
Some tests were not using a copy of the global environment and were modifying os.environ. This is a bad practice in the tests. While the fix is easy, preventing this from happening again is not possible with the current code since we want it to be easy to use the global environment.