tests: Use copy of env for init (#6513)
Some tests were not using a copy of the global environment and were modifying os.environ. This is a bad practice in the tests. While the fix is easy, preventing this from happening again is not possible with the current code since we want it to be easy to use the global environment.

The global environment is still used in tests of grass.jupyter, but there we need first to add an env parameter everywhere to enable the isolated testing, so that's not covered in this changeset.