CQ: Fix Ruff PLR6201: Use a set literal when testing for membership
Ruff rule: https://docs.astral.sh/ruff/rules/literal-membership

See the note in the docs: The left and the right hand side must be hashable for it to not throw an error at runtime. That's why when I was unsure of what could be passed, I ignore the rule for the file.