CQ: we should check the return value of lseek (#5783)
An attempt to unify checks and error messages.

There are more places where `lseek` return value should be checked, but I don't have time to work on them now.

* CQ: we should check the return value of lseek
An attempt to unify checks and error messages

* r.random: remove unnecesarry lseek

Rewinding a raster map can lead to lseek errors and generally
should not be done, as raster reading functions perform seek
on their own if required.

* QA: add more checks for lseek return value

* QC: finish cleaning up lseek error handling:

* no need to cast -1 to off_t as we do not use K&R C
* make error message arguments relocable
* unify error messges
* add missing includes for error reporting

* Fix wording of error message on seek fail

* Clean up unintentional changes