r.stream.extract: restore removed code making r.stream.extract more robust (#6543)
This block got removed as part of compiler warnings cleanup #1248, but the removal broke special cases (#6541).
The intention of r.stream.extract is to use external flow accumulation for stream extraction, but if in doubt follow the A* search direction, i.e. if (fabs(wat_nbr[np_side]) >= max_acc).
Restoring this deleted code block in addition to #6554 should make r.stream.extract robust for special cases as reported in #6541.