Add nprocs to r.blend
The main step in r.blend is running r.mapcalc which is now multithreaded. The default number of threads is now whatever is available on the machine, so to limit that, r.blend needs nprocs which is passed to r.mapcalc (before that change, r.blend was not taking the advantage of multithreading in r.mapcalc).