grass.pygrass: Set r.mapcalc threads in GridModule
In the serial portions of the pygrass GridModule processing, use the number of processing elements passed to GridModule as the nprocs parameter of r.mapcalc. This is used when rasters are copied over and r.mapcalc is used to create a more suitable version of the raster. (There might be a better way of implementing that step, but with the current code, using the right number of cores/threads is the obvious improvement and fix.)

When the code already runs in parallel, set number of processing elements to 1 so that we avoid oversubscription within the process itself.