r.grow: Expose r.mapcalc's nprocs
One of the two steps in r.grow is using r.mapcalc. This allows user to control the number of cores used by r.mapcalc in that step (#5742). This is more critical with the default being all cores (#5731) even if the impact is possibly limited given that the main step is not parallelized.