grass.pygrass: Set r.mapcalc nprocs in ParallelModuleQueue example (#6583)
Example of ParallelModuleQueue from pygrass uses r.mapcalc. With r.mapcalc being parallel itself, we need to deal with number of processing units for the r.mapcalc tool itself, especially when the default currently is all cores. The new example presents the advanced, but realistic, scenario when we combine the two parallelizations together. Readers can always go with the hopefully obvious solution of setting nprocs=1, so having a more complete example in place seems justified as it covers the full spectrum of choices.