r.mapcalc.simple: Add nprocs for r.mapcalc (#6584)
With r.mapcalc being parallel (#5742) and the default nprocs in C (for r.mapcalc) being all cores (#5731), it both makes sense and is necessary to have a pass-through nprocs in r.mapcalc.simple. This gives users access to the parallelization options (regardless of the default), and it allows users to avoid oversubscription (with the default being all cores).

This depends on #6580 for the nprocs parameter in the gs.mapcalc function.