i.image.mosaic: Allow to set number of cores (#6586)
The i.image.mosaic tool uses r.mapcalc for the main processing step. This allows user to control the number of cores used by r.mapcalc (since OpenMP parallelization was added to r.mapcalc in #5742). With the default number of cores used for processing being all cores (#5731), the parallelization is enabled by default. This allows users to control number of cores used allowing for specific number of cores to be used.

This depends on #6580 for the nprocs parameter in the gs.mapcalc function.