make: pass GDAL include dirs to proj
vector_inc and vedit_inc don't need this treatment as VECT_CFLAGS has it
added already. This would conform to how its done in the CMakeLists.txt
as well.

Oddly enough this isn't a fatal error though. So if anyone hit a build
failure and thought this was the issue should look further into their
builds logs for the actual cause.

Bug: https://github.com/OSGeo/grass/issues/4258
Signed-off-by: Alfred Wingate <parona@protonmail.com>