grass.app: Always set path to script addons
For relevant platforms (currently all except Windows), add the scripts directory for addons to path even if the directory does not exist. When installing into a fresh install, the directory may not exist, so it is not added and the addon tool does not run.

The new test fails for the Python tool with the old code, but works with the new code.